Fix bug with initialization of costs matrix.
Ensure that the minimum cost operation is chosen to initialize the
costs matrix instead of only choosing deletion as the default activity.

Using deletion as the default activity works in the case insertion_cost
and deletion_costs are the same. However, if they are different, then
the initialization method makes a difference to the final result.